View comparison optimization (#26)
* PawnList8

* Prog...

* Partial

* Works for d6

* Add impl for non-sse

* d3 v

* k4 e

* Remaining

* fuzz_equals_ignoring_order

* Sort pawns progress

* Fuzz sort test

* Impl but broke

* Fixed sort

* Test inequal views

* Test equal ignoring hash

* PackedIdx::maybe_from

* Impld but borke

* Handle null pawn indexes in pawns list

* Use lookup table for board symm state for N = 16

* Improve equal_ignoring_order_sse by not sorting

* Cleanup + unreachable

* Better profiling

* Update gitignore

* Super fast apply_d6_c_sse

* Maybe table-driven apply?

* Apply normalizing op to both sets of pawns

* Cleanup eq_ignoring_order

* Inline board_symm_state

* Profile script

* Special case origin for pawns_in_play == N

* Make BoardSymmetryState smaller

* Cache pawns_in_play

* Separate two impls of state comparison in onoro view

* Reorder some code, add comments.

* Comment.

* Rm todo

* Pawns list explain rotate_impl

* Comments

* Comments

* sse4.1 -> ssse3

* Comments.

* Move

* Comments

* Comments

* Debug check

* Fix null index handling in PawnList8

* Add cfg condition to specialized view comparisons

+++ onoro_impl/src/canonical_view.rs
@@ -41,8 +41,8 @@ impl CanonicalView {
     self.symm_class
   }
 
-  pub fn op_ord(&self) -> u8 {
-    self.op_ord
+  pub fn op_ord(&self) -> usize {
+    self.op_ord as usize
   }
 
   pub fn hash(&self) -> u64 {
